折り方を決定するカスタムフック
